// src/pages/interfaces/comercio.ts
export interface Comercio {
    logo_url: string | undefined;
    id: number;
    ruc: string;
    nombre: string;
    descripcion: string;
    direccion: string;
    email: string;
    contacto_nombre_apellido: string;
    contacto_celular: string;
    contacto_email: string;
    fecha_registro: string;
}